Split client file
